[project @ 2004-01-12 06:56:05 by craigmcc] [JSFA148]
Commit the rest of the patch that Hans originally proposed, along with unit
test updates to prove that updating the model works in either Apply Request
Values or Invoke Application phase.  This gives us a restart point for
evaluating refinements.

However, making this work in the "UIData" page of the "standard" package in
jsf-demo pointed out that an application action needs to know which phase
it was invoked in -- if it's Appy Request Values (because of immediate=true
on the UICommand), then you can no longer look at the "value" property of a
per-row input component; you can only look at the "submittedValue" property.
And that's a problem, because what is stored here is an internal implementation
detail of the Renderer.  But at least it works in both immediate modes
at the moment.

Original author: craigmcc
Date: 2004-01-12 06:56:07+00:00