Fix for issue 1020 - Components inside f:ajax removed on postback
The fix is localized in AjaxHandler.apply().  The problem is that AjaxHandler.apply() was not calling through to nextHandler.apply() on postbacks (ie. when !ComponentHandler.isNew(parent)).  As a result, children of the <f:ajax> are not re-applied, and end up being orphaned/removed from the component tree.

Unfortunately don't have time to add an automated test at the moment - though as Ryan suggested in the the issue detail, we clearly need to add one.  For the moment, the behavior can be verified manually in the ajax-tag demo.

r=rogerk