Bug 1307 - we were emitting invalid XML into the page, since partial responses were wrapped in CDATA, and were thus not escaped.
This is a failing of the specification, and I've filed spec bug 637 to track
it.  

Note that this fix is liable to be somewhat fragile - the handling of doctype and xml declarations in particular.

SECTION: Modified Files
----------------------------
M       jsf-ri/src/com/sun/faces/context/PartialViewContextImpl.java

-- Set the underlying HTMLResponseWriter to be aware that it's in a partial
response.  This is a fragile solution, but will probably not cause too many 
problems.

M       jsf-ri/src/com/sun/faces/renderkit/html_basic/HtmlResponseWriter.java

-- Add isPartial() method, then do appropriate escaping if we're in a partial
response.  Unfortunately, this hoses the xml declaration and 
doctype declaration, so a special exception is necessary. That exception is
going to be extremely fragile, but we'll have to address this in the next 
version of the spec.  If I could think of a better way of doing this...

M       jsf-ri/systest/src/com/sun/faces/ajax/AjaxErrorTestCase.java

-- Error messages will now have their xml entities escaped.  This will mean, 
primarily, that " will become &quot; - while not pretty, I could not think of
any way around this.